tool: move custom builders into their own file
Hopefully towards using build.yaml for this package
